backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.1
pydantic==2.10.4
pydantic-settings==2.7.1
celery[redis]==5.4.0
redis==5.2.1
python-dotenv==1.0.1
httpx==0.28.1
cryptography==44.0.0
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
bcrypt==5.0.0
pytest==8.3.4
pytest-asyncio==0.24.0

# E-mail
sendgrid==6.11.0

# Google Calendar
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-api-python-client==2.163.0
